Fix lighting problem
The problem's cuase is because we use differrent space coordinate in the fragment shader. That is NormalMatrix in view space, Position and Direction in world space.
Finally I add a view matrix calculation on CELight's updateUniformsWithCamera: method.
By the way, I change CEObject.haschanged property, now it also indicates the change of its parent's.